Added Linked List directory & README.md
