--- C++/Abstract DT/Linked List.cpp
@@ -1,38 +1,0 @@
-#include <cstddef>
-#include <iostream>
-
-using namespace std;
-
-class Node {
-    public:
-        int data;
-    Node * next;
-};
-
-void print_list(Node * n) {
-    while (n != NULL) {
-        cout << n->data << " ";
-        n = n->next;
-    }
-}
-
-int main() {
-    Node * head = NULL;
-    Node * second = NULL;
-    Node * third = NULL;
-
-    head = new Node();
-    second = new Node();
-    third = new Node();
-
-    head->data = 1;
-    head->next = second;
-
-    second->data = 2;
-    second->next = third;
-
-    third->data = 3;
-    third->next = NULL;
-
-    print_list(head);
-}
