+++ C++/Abstract DT/Linked List/Linked List.cpp
@@ -1,0 +1,198 @@
+//Declaration
+#include <cstddef>
+#include <iostream>
+
+using namespace std;
+
+class Node {
+    public:
+        int data;
+    Node * next;
+};
+
+void print_list(Node * n) {
+    while (n != NULL) {
+        cout << n->data << " ";
+        n = n->next;
+    }
+}
+
+int main() {
+    Node * head = NULL;
+    Node * second = NULL;
+    Node * third = NULL;
+
+    head = new Node();
+    second = new Node();
+    third = new Node();
+
+    head->data = 1;
+    head->next = second;
+
+    second->data = 2;
+    second->next = third;
+
+    third->data = 3;
+    third->next = NULL;
+
+    print_list(head);
+}
+
+
+//Inserting Node at start
+#include <cstddef>
+#include <iostream>
+
+using namespace std;
+
+class Node {
+    public:
+        int data;
+    Node * next;
+};
+
+void print_list(Node * n) {
+    cout << "\nPrinting new list..." << endl;
+    while (n != NULL) {
+        cout << n->data << " ";
+        n = n->next;
+    }
+} 
+
+void push(struct Node ** head_ref, int new_data) {
+    struct Node * new_node = (struct Node * ) malloc(sizeof(struct Node));
+    new_node->data = new_data;
+    new_node->next = ( * head_ref);
+    ( * head_ref) = new_node;
+} 
+
+int main() {
+    Node * head = NULL;
+    Node * second = NULL;
+    Node * third = NULL;
+
+    head = new Node();
+    second = new Node();
+    third = new Node();
+
+    head->data = 1;
+    head->next = second;
+
+    second->data = 2;
+    second->next = third;
+
+    third->data = 3;
+    third->next = NULL;
+
+    print_list(head);
+    push(&head, 11);
+    print_list(head);
+}
+
+
+//Inserting Node at end
+#include <cstddef>
+#include <iostream>
+
+using namespace std;
+
+class Node {
+    public:
+        int data;
+    Node * next;
+};
+
+void print_list(Node * n) {
+    cout << "\nPrinting new list..." << endl;
+    while (n != NULL) {
+        cout << n->data << " ";
+        n = n->next;
+    }
+} 
+
+void append(struct Node ** head_ref, int new_data) {
+    struct Node * new_node = (struct Node * ) malloc(sizeof(struct Node));
+    struct Node * last = * head_ref;
+    new_node->data = new_data;
+    new_node->next = NULL;
+    while (last->next != NULL)
+        last = last->next;
+    last->next = new_node;
+}
+
+int main() {
+    Node * head = NULL;
+    Node * second = NULL;
+    Node * third = NULL;
+
+    head = new Node();
+    second = new Node();
+    third = new Node();
+
+    head->data = 1;
+    head->next = second;
+
+    second->data = 2;
+    second->next = third;
+
+    third->data = 3;
+    third->next = NULL;
+
+    print_list(head);
+    append(&head, 11);
+    print_list(head);
+}
+
+//Inserting Node at specific position
+#include <cstddef>
+#include <iostream>
+
+using namespace std;
+
+class Node {
+    public:
+        int data;
+    Node * next;
+};
+
+void print_list(Node * n) {
+    cout << "\nPrinting new list..." << endl;
+    while (n != NULL) {
+        cout << n->data << " ";
+        n = n->next;
+    }
+} 
+
+void insertAfter(struct Node * prev_node, int new_data) {
+    if (prev_node == NULL) {
+        printf("the given previous node cannot be NULL");
+        return;
+    }
+    struct Node * new_node = (struct Node * ) malloc(sizeof(struct Node));
+    new_node->data = new_data;
+    new_node->next = prev_node->next;
+    prev_node->next = new_node;
+}
+
+int main() {
+    Node * head = NULL;
+    Node * second = NULL;
+    Node * third = NULL;
+
+    head = new Node();
+    second = new Node();
+    third = new Node();
+
+    head->data = 1;
+    head->next = second;
+
+    second->data = 2;
+    second->next = third;
+
+    third->data = 3;
+    third->next = NULL;
+
+    print_list(head);
+    insertAfter(second, 11);
+    print_list(head);
+}
