+++ C++/Abstract DT/Linked List/README.md
@@ -1,0 +1,10 @@
+Author : ITES7321
+
+Added:
+-Implementation of Linked list
+-Implementation of algo of adding nodes to Linked list at start,end or any given specific position
+
+Changes:
+-Added Abstract DT directory + Linked List directory
+
+Note:Linked List.cpp is well supported with comments to help readability.
